requirements.txt
black
coverage
librosa
matplotlib
nnAudio
PyYAML
tensorboard
torch
torchaudio
